% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_files.R
\name{analyze.files}
\alias{analyze.files}
\title{Analyze your filesystem.}
\usage{
analyze.files(path = getwd(), all.files = TRUE, ...)
}
\arguments{
\item{path}{a path to analyse in your filesystem.}
}
\value{
A tibble
}
\description{
Analyze your filesystem.
}
\examples{
analyze.files() \%>\% filter(!grepl("Rproj", path))
}
