% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_folder.R
\name{build_folder}
\alias{build_folder}
\title{build R proyect folders}
\usage{
build_folder(rmd = FALSE, dir_tree = TRUE)
}
\arguments{
\item{rmd}{if is TRUE an Rmds folder will be created.}

\item{dir_tree}{if is TRUE an directory tree will be presented in the console.}
}
\description{
build R proyect folders
}
